tests: Move encoding tests to ogorek_test.go
We are going to integrate encoding tests with the main tests data table
in several steps.

Only code movement here, no other change.
